[BEAM-11188] Refactor Go xlang expansion from runners to core.
Moving the xlangx calls out of universal.go and into graphx/translate.go caused a dependency loop (graphx -> xlangx -> graphx), so in order to break the cycle I moved all xlangx calls that were used in graphx directly into graphx, so now it looks like this: xlangx -> graphx.

Most of those calls were in translate.go, so I just moved the contents of that file into graphx/xlang.go, which is a sensible place for it anyway since it's a lot of code modifying proto pipelines.
